+++ editor/src/main/java/com/github/inc0grepoz/kvad/entities/being/Being.java
@@ -1,13 +1,13 @@
 package com.github.inc0grepoz.kvad.entities.being;
 
-import java.awt.Dimension;
 import java.awt.Graphics;
-import java.awt.Rectangle;
 import java.awt.image.BufferedImage;
 
-import com.github.inc0grepoz.kvad.Vector;
+import com.github.inc0grepoz.kvad.awt.geom.Dimension;
+import com.github.inc0grepoz.kvad.awt.geom.Rectangle;
 import com.github.inc0grepoz.kvad.entities.Renderable;
 import com.github.inc0grepoz.kvad.entities.level.Level;
+import com.github.inc0grepoz.kvad.utils.Vector;
 
 import lombok.Getter;
 
@@ -92,7 +92,7 @@ public class Being extends Renderable {
 
     @Override
     public void typeText(Graphics gfx, Rectangle cam, Rectangle rect) {
-        int x = (int) (rect.getCenterX() - cam.x), y = rect.y + rect.height - cam.y;
+        int x = (int) (rect.getCenterX() - cam.x), y = (int) (rect.y + rect.height - cam.y);
         int width = gfx.getFontMetrics().stringWidth(type);
         x -= width / 2;
         y += 15;
